% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_data_fct_display.R
\name{compare_data_frame_object}
\alias{compare_data_frame_object}
\title{Function which perform the comparison of dataframe}
\usage{
compare_data_frame_object(df1, df2, id_var)
}
\arguments{
\item{df1}{The first dataframe of the comparison}

\item{df2}{The second dataframe of the comparison}

\item{id_var}{Tve vector of id variable to identify the observations}
}
\description{
Function which perform the comparison of dataframe
}
