% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCompare-package.R
\docType{package}
\name{dataCompare-package}
\alias{dataCompare}
\alias{dataCompare-package}
\title{dataCompare: data Comparison}
\description{
Application to compare two data frame and provide differences.
}
\author{
\strong{Maintainer}: Sergio Ewane \email{info@dataforknow.com}

}
\keyword{internal}
