% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_add_external_resources.R
\name{golem_add_external_resources}
\alias{golem_add_external_resources}
\title{Add external Resources to the Application}
\usage{
golem_add_external_resources()
}
\value{
No return value, called for side effects
}
\description{
This function is internally used to add external
resources inside the Shiny application.
}
