% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_comp_details.R
\name{mod_comp_details_ui}
\alias{mod_comp_details_ui}
\title{mod_comp_details_ui and mod_comp_details_server}
\usage{
mod_comp_details_ui(id)
}
\arguments{
\item{id}{an id}
}
\value{
No return value
}
\description{
A shiny module.
}
