% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_data_compare_app}
\alias{run_data_compare_app}
\title{Run the dataCompare Shiny Application}
\usage{
run_data_compare_app(...)
}
\arguments{
\item{...}{list of arguments}
}
\value{
No return value, launch the app
}
\description{
Run the dataCompare Shiny Application
}
