% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_data_fct_display.R
\name{same_variables}
\alias{same_variables}
\title{Detect common variables in two dataset}
\usage{
same_variables(df1, df2)
}
\arguments{
\item{df1}{the first dataset to use}

\item{df2}{The second dataset to use}
}
\description{
Detect common variables in two dataset
}
