% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_data_fct_display.R
\name{skim_char}
\alias{skim_char}
\title{Skim a dataset and return only characters variables characteristics}
\usage{
skim_char(the_data)
}
\arguments{
\item{the_data}{Data on which the skim function will apply the description on character variables}
}
\description{
Skim a dataset and return only characters variables characteristics
}
