% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCompare-package.R
\docType{package}
\name{dataCompare-package}
\alias{dataCompare}
\alias{dataCompare-package}
\title{dataCompare: Data Comparison}
\description{
A shiny app use to compare two data frame and provide more details on differences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seewe/dataCompare}
  \item Report bugs at \url{https://github.com/seewe/dataCompare/issues}
}

}
\author{
\strong{Maintainer}: Sergio Ewane Ebouele \email{info@dataforknow.com}

}
\keyword{internal}
