% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCompare-package.R
\docType{package}
\name{dataCompare-package}
\alias{dataCompare}
\alias{dataCompare-package}
\title{dataCompare: A 'shiny' app to compare two data frames}
\description{
The 'dataCompare' package is a 'shiny' application developed with the 'Golem' framework. It provides a graphical tool for checking value differences between two data frames. The user provides two data frames for comparison, selects IDs variables identifying each row of input data, then clicks a button to perform the comparison. Several 'R' package functions are used to describe the data and perform the comparison in the backend of the application. The main ones are 'comparedf()' from 'arsenal' and 'skim()' from 'skimr'. See the description of 'comparedf()' from the 'arsenal' package here \url{https://cran.r-project.org/web/packages/arsenal/vignettes/comparedf.html} and that of 'skim()' from the 'skimr' package here \url{https://cran.r-project.org/web/packages/skimr/vignettes/skimr.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seewe/dataCompare}
  \item Report bugs at \url{https://github.com/seewe/dataCompare/issues}
}

}
\author{
\strong{Maintainer}: Sergio Ewane Ebouele \email{info@dataforknow.com}

}
\keyword{internal}
